Render tags and links to SauceLabs sessions
Tags are displayed for all features and feature elements. Clicking a tag
results in filtering of the results by that tag value.

Embedded data with the 'text/url' MIME type are now rendered as icons
and external links. This currently applies to SauceLabs sessions.

Error messages are also displayed now, collapsed by default and expanded
upon clicking.

Riot dependency was upgraded for support of references to nested tag
collections (via `this.tags[...]`).

Change-Id: I71b6fd2eac90db3eea36bb1a708a16c1cc30c5a3